Supply open() implementation for boto store which supports seek and tell.
