+++ tests/test_azure_store.py
@@ -2,7 +2,7 @@ from uuid import uuid4 as uuid
 from simplekv._compat import ConfigParser, pickle
 from simplekv.net.azurestore import AzureBlockBlobStore
 from simplekv.contrib import ExtendedKeyspaceMixin
-from basic_store import BasicStore
+from basic_store import BasicStore, OpenSeekTellStore
 from conftest import ExtendedKeyspaceTests
 import pytest
 
@@ -36,7 +36,7 @@ def create_azure_conn_string(credentials):
     return fmt_str.format(account_name, account_key)
 
 
-class TestAzureStorage(BasicStore):
+class TestAzureStorage(BasicStore, OpenSeekTellStore):
     @pytest.fixture
     def store(self):
         from azure.storage.blob import BlockBlobService
@@ -48,60 +48,6 @@ class TestAzureStorage(BasicStore):
         yield AzureBlockBlobStore(conn_string=conn_string, container=container,
                                   public=False)
         s.delete_container(container)
-
-    def test_open_seek_and_tell_empty_value(self, store, key):
-        value = b''
-        store.put(key, value)
-        ok = store.open(key)
-        assert ok.seekable()
-        ok.seek(10)
-        assert ok.tell() == 10
-        ok.seek(-6, 1)
-        assert ok.tell() == 4
-        with pytest.raises(IOError):
-            ok.seek(-1, 0)
-        with pytest.raises(IOError):
-            ok.seek(-6, 1)
-        with pytest.raises(IOError):
-            ok.seek(-1, 2)
-
-        assert ok.tell() == 4
-        assert b'' == ok.read(1)
-
-    def test_open_seek_and_tell(self, store, key, long_value):
-        store.put(key, long_value)
-        ok = store.open(key)
-        assert ok.seekable()
-        assert ok.readable()
-        ok.seek(10)
-        assert ok.tell() == 10
-        ok.seek(-6, 1)
-        assert ok.tell() == 4
-        with pytest.raises(IOError):
-            ok.seek(-1, 0)
-        with pytest.raises(IOError):
-            ok.seek(-6, 1)
-        with pytest.raises(IOError):
-            ok.seek(-len(long_value) - 1, 2)
-
-        assert ok.tell() == 4
-        assert long_value[4:5] == ok.read(1)
-        assert ok.tell() == 5
-        ok.seek(-1, 2)
-        length_lv = len(long_value)
-        assert long_value[length_lv - 1:length_lv] == ok.read(1)
-        assert ok.tell() == length_lv
-        ok.seek(length_lv + 10, 0)
-        assert ok.tell() == length_lv + 10
-        assert b'' == ok.read()
-
-        ok.close()
-        with pytest.raises(ValueError):
-            ok.tell()
-        with pytest.raises(ValueError):
-            ok.read(1)
-        with pytest.raises(ValueError):
-            ok.seek(10)
 
 
 class TestExtendedKeysAzureStorage(TestAzureStorage, ExtendedKeyspaceTests):
